refactor(cron): 将 useFormatNextRunRelative hook 提取为纯函数 formatNextRunRelative
该函数仅依赖 dayjs 和 t()，无需 useState/useEffect 等 hook 能力。
提取到 utils 后可在组件中直接调用，消除不必要的 hook 封装。
同时将返回按钮图标从 ArrowLeft 换为 Left，增加 hover 态和点击热区。
